+++ rp-api/api/src/gen_handler.h
@@ -57,8 +57,8 @@ int gen_setBurstCount(rp_channel_t channel, int num);
 int gen_getBurstCount(rp_channel_t channel, int* num);
 int gen_setBurstRepetitions(rp_channel_t channel, int repetitions);
 int gen_getBurstRepetitions(rp_channel_t channel, int* repetitions);
-int gen_setBurstPeriod(rp_channel_t channel, uint32_t period);
-int gen_getBurstPeriod(rp_channel_t channel, uint32_t* period);
+int gen_setBurstPeriod(rp_channel_t channel, float period);
+int gen_getBurstPeriod(rp_channel_t channel, float* period);
 
 int gen_setLoadMode(rp_channel_t channel, rp_gen_load_mode_t mode);
 int gen_getLoadMode(rp_channel_t channel, rp_gen_load_mode_t* mode);
